+++ src/tag/tag.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { Tag } from '../entities/tag.entity';
-import { CreateTagRequestDto } from './dto/request-tag.dto';
+import { CreateTagRequestDto } from './dto/request/request-tag.dto';
 import { TagRepository } from './repository/tag.repository';
 
 @Injectable()
